refactor: 비즈니스에 따른 도메인 경계 분리
비즈니스에 맞도록 도메인 경계 분리

Ref: #2
