+++ src/main/java/lotto/model/lotto/Player.java
@@ -1,4 +1,4 @@
-package lotto.model;
+package lotto.model.lotto;
 
 import java.util.List;
 import java.util.stream.Collectors;
